Prototype v4.1
Additions and Changes:
- Added Health Bar UI and system to Arcadian Controller (to be replicated on all Turrets)
- Added AttackPoint to Arcadian Controller for Enemy Targeting (to be replicated on all Turrets)
- Changed Enemy Attack/Movement system to create better AI
- Changed Turret and Enemy Health stat from INT to FLOAT
- Attack Range balance on all enemies and turrets
- Fixed bug where waves would continue to spawn after Game Over state
- Fixed bug where level win screen could still show after Game Over state
- Fixed bug where Enemy health bar was not correctly displaying lost health amount

+++ Assets/Scripts/EnemyUnits/EnemyStats.cs
@@ -9,8 +9,8 @@ public class EnemyStats
     public int normalValue;
     public int specialValue;
 
-    public int startHealth;
-    public int health;
+    public float startHealth;
+    public float health;
     public int damage;
 
     public float startSpeed;
